+++ packages/docz-core/src/Bundler.ts
@@ -20,8 +20,6 @@ export interface BundlerConstructor<Config> {
   build: BuildFn<Config>
 }
 
-const IS_PROD = process.env.NODE_ENV === 'production'
-
 export class Bundler<C = any> {
   private readonly args: Args
   private config: C
@@ -38,8 +36,8 @@ export class Bundler<C = any> {
   }
 
   public getConfig(): C {
-    const config = this.args.modifyBundlerConfig(this.config, !IS_PROD)
-    return this.mountConfig(config)
+    const config = this.mountConfig(this.config)
+    return this.args.modifyBundlerConfig(config, !this.isProd())
   }
 
   public async createServer(config: C): Promise<BundlerServer> {
@@ -54,6 +52,10 @@ export class Bundler<C = any> {
     const { plugins } = this.args
     const reduce = Plugin.reduceFromPlugins<C>(plugins)
 
-    return reduce('modifyBundlerConfig', config, !IS_PROD)
+    return reduce('modifyBundlerConfig', config, !this.isProd())
+  }
+
+  private isProd(): boolean {
+    return process.env.NODE_ENV === 'production'
   }
 }