+++ packages/docz-core/src/utils/load-config.ts
@@ -4,6 +4,7 @@ import * as paths from '../config/paths'
 import { Config } from '../commands/args'
 import { Plugin } from '../Plugin'
 import { omit } from './helpers'
+import { BabelRC } from './babelrc'
 
 const toOmit = ['_', '$0', 'version', 'help']
 
@@ -11,11 +12,13 @@ export const loadConfig = (args: Config): Config => {
   const config = load('docz', {
     ...args,
     paths,
+    hashRouter: false,
     plugins: [],
     mdPlugins: [],
     hastPlugins: [],
     themeConfig: {},
     modifyBundlerConfig: (config: any) => config,
+    modifyBabelRc: (babelrc: BabelRC) => babelrc,
   })
 
   const reduce = Plugin.reduceFromPlugins<Config>(config.plugins)