+++ packages/docz-plugin-css/librc.js
@@ -1,0 +1,7 @@
+const pkg = require('./package.json')
+
+module.exports = {
+  external: Object.keys(pkg.dependencies).concat([
+    'react-dev-utils/getCSSModuleLocalIdent',
+  ]),
+}