+++ packages/docz-theme-default/src/components/shared/Main/index.tsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import { Component } from 'react'
 import styled from 'react-emotion'
 import equals from 'fast-deep-equal'
-
 import { base } from '../../../styles/base'
+import webfont from 'webfontloader'
 
 const Wrapper = styled('div')`
   display: flex;
@@ -13,6 +13,13 @@ const Wrapper = styled('div')`
 interface MainProps {
   config: any
 }
+
+const loadfonts = () =>
+  webfont.load({
+    google: {
+      families: ['Inconsolata', 'Source Sans Pro:300,400,600,700'],
+    },
+  })
 
 export class Main extends Component<MainProps> {
   public componentDidUpdate(prevProps: MainProps): void {
@@ -25,6 +32,7 @@ export class Main extends Component<MainProps> {
 
   public componentDidMount(): void {
     base(this.props.config)
+    loadfonts()
   }
 
   public render(): React.ReactNode {