+++ packages/docz/librc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  commonjs: {
+    namedExports: {
+      '../../node_modules/@mdx-js/tag/dist/index.js': ['MDXProvider'],
+    },
+  },
+}