+++ packages/docz/src/types.d.ts
@@ -1,2 +1,4 @@
 declare module '@sindresorhus/slugify'
+declare module '@mdx-js/tag'
+declare module '@mdx-js/tag/dist/mdx-provider'
 declare module 'pascalcase'